This is a Lua script written for Neovim, a popular and highly configurable text editor. The script defines the configuration for a plugin called `no-neck-pain.nvim`, which aims to prevent eye strain caused by splitting the Neovim window horizontally or vertically.
The configuration allows users to set options for the left and right buffers, as well as buffer-scoped options that apply to both side buffers. Users can also configure various integrations, such as NvimTree and NeoTree, to ensure proper positioning of the plugin with these integrations.

The main function `setup` is used to initialize the plugin by setting up the configuration options and creating a custom autocommand that automatically enables/disables the plugin when opening or closing buffers. The plugin also provides an optional statusline icon for easy enabling/disabling in the statusbar.